fix 32 bit build

+++ include/cextras/collection.h
@@ -579,7 +579,7 @@ struct CxRadixTree {
 	struct CxRadixNode *root;
 };
 
-void cx_radix_tree_init(struct CxRadixTree *map, uint64_t element_size);
+void cx_radix_tree_init(struct CxRadixTree *map, size_t element_size);
 
 void *cx_radix_tree_get(const struct CxRadixTree *map, uint64_t key);
